% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covgccak.R
\name{covgccak}
\alias{covgccak}
\title{The function covgccak() is called by fgcca() and does not have to be used by
the user. The function returns the canonical functions/vectors for random object}
\usage{
covgccak(
  megaCov,
  grids,
  C,
  W,
  tau = rep(1, length(megaCov)),
  scheme = "centroid",
  verbose = FALSE,
  init = "random",
  tol = 1e-16
)
}
\arguments{
\item{megaCov}{A list of lists containing the covariance/cross-covariance matrices}

\item{grids}{The work grid for each process}

\item{C}{The design matrix of the FGCCA model (J*J)}

\item{W}{A list of weighting matrices, allowing for functional data to perfom numerical integration}

\item{tau}{The regularization parameter for each process}

\item{scheme}{The scheme function (should be 'horst', 'factorial' or 'centroid')}

\item{verbose}{A logical value indicating if verbose}

\item{init}{The initialization method (should be 'random' or 'svd')}

\item{tol}{Tolerance parameter for the algorithm}
}
\value{
fgcca object#'
}
\description{
The function covgccak() is called by fgcca() and does not have to be used by
the user. The function returns the canonical functions/vectors for random object
}
