% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcca.R
\name{fgcca}
\alias{fgcca}
\title{Functional Generalized Canonical Correlation Analysis (FGCCA)}
\usage{
fgcca(
  Lys,
  Lts,
  connection = 1 - diag(length(Lys)),
  tau = rep(1, length(Lys)),
  scheme = "factorial",
  init = "random",
  tol = 1e-14,
  deflType = "ortho",
  verbose = TRUE,
  center = TRUE,
  scale = FALSE,
  ncomp = rep(1, length(Lys)),
  optns = list()
)
}
\arguments{
\item{Lys}{A J list of n lists containing the observations}

\item{Lts}{A J list of n lists containing the time-points}

\item{connection}{The design matrix of the FGCCA model (J*J)}

\item{tau}{The regularization parameter for each process}

\item{scheme}{The scheme function (should be 'horst', 'factorial' or 'centroid')}

\item{init}{The initialization method (should be 'random' or 'svd')}

\item{tol}{Tolerance parameter for the algorithm}

\item{deflType}{The deflation type (should be 'uncor' (default) or 'ortho')}

\item{verbose}{A logical value indicating if showing computation details}

\item{center}{A logical value indicating if applying centering on the processes}

\item{scale}{A logical value indicating if applying a variance scaling on the processes}

\item{ncomp}{The number of components to extract for each process}

\item{optns}{A list of options}
}
\value{
A fgcca object containing the canonical functions (a) and the canonical components (Y)
}
\description{
Functional Generalized Canonical Correlation Analysis (FGCCA) is a flexible framework for exploring relationships between any number J of random processes.
It is robust to sparse and irregular functional data.
}
\details{
Options for FGCCA are :
\describe{
Smoothing and score computation options (optns)
\item{smoothing}{"gam", "lwls" or "ns" (no smoothing), must be on a regular grid. Default is "gam".}
\item{gridSize}{Size of the smoothing grids. Default is 50.}
\item{scoreConnection}{Connection matrix used in the Gaussian conditioning. Default is full of 1 matrix. Can be useful to change for supervised learning}
\item{scoreMethod}{"CE" or "IN". "CE" is Conditional Expectation (Bayesian Approach). "IN" is by numerical INtegration Default is "CE".}
}
}
